v0.4 FEATURE COMPLETE!!! (#39)
* UI Improvements (#31)

- Added collapsible camera menus
- Added beginnings of event filtering
  - Can now select number of events without opening the menu

* V0.4 dev (#33)

* UI Improvements

- Added collapsible camera menus
- Added beginnings of event filtering
  - Can now select number of events without opening the menu

* Documentation Cleanup

Update README.md
Clean up template.env

* update rest.py

update access to sqlite class from within rest.py.

* Add filtering options to events page

* Fixes for Issues #14  and #15

- #14:
  - added regex check to see if its a valid event id.  If not, it will not delete.
- #15:
  - value must equal 'debug' 'info' or 'error' otherwise it will not proceed.

* Filters, Settings, and UI changes

- Added functional filters that update the page when the value is changed.
- Added option to choose 12/24 hour clock
- Tweaked some design changes.

* Event Filtering

Added "Pager" to navigate pages.
Updated filters and related bugfixes.

* Update event.py

* Bugfixes

Fixed several bugs in filter.py and events.py.
- Changed how required query strings are parsed

+++ docker/fEVR/rootfs/var/www/html/cgi-bin/fetch.py
@@ -38,7 +38,11 @@ class fetchEvent:
     def delEvent(self,db=True):
         import shutil
         path = f"{self.location}{self.event}"
-        shutil.rmtree(path)
+        import re
+        if re.match(r'\d{10}.\d{5}-\w{6}',self.event):
+            shutil.rmtree(path)
+        else:
+            self.error.execute(f"delEvent(): {self.event} is not a valid event.  Someone is trying to do something silly silly!!!",src=self.script)
         if db:
             self.error.execute(f" delEvent() - DELETING DATABASE ENTRY",src=self.script)
             sql = f"""DELETE FROM events WHERE event='{self.event}';"""
